Move binarysearch list extensions and add comment to indicate the source.
